# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: AI Homework Grader API
  version: "0.2.0"
  description: |
    API for AI-powered homework grading. Upload exam photos, receive annotated
    grading results, and browse error-question collections. MVP covers English
    and Math.

    ## Authentication (v0.2.0)
    Parent identity is established via SMS verification-code login
    (`POST /api/auth/send-code` + `POST /api/auth/login`), which issues a JWT.
    All business endpoints require `Authorization: Bearer <token>`; the JWT
    `sub` claim is the parent id (userId). Phone appears ONLY in the auth
    endpoints — never in business URLs, query params, or logs.

    Image endpoints use signed URLs (HMAC token + expiry in the query string)
    because `<img>`/`<image>` tags cannot send an Authorization header.

    The WeChat mini-program uses the SAME SMS login flow as the Web client
    (`POST /api/auth/send-code` + `POST /api/auth/login`); there is no
    mini-program-specific login endpoint. The issued JWT is cached client-side
    and reused until it expires.
servers:
  - url: http://localhost:8000
    description: Local development (FastAPI backend)

# Global security: all business endpoints require a Bearer JWT unless overridden.
security:
  - bearerAuth: []
...
paths:
...
  /api/submissions:
    post:
      operationId: createSubmission
      tags: [Submissions]
      summary: Upload an exam image for async grading
      description: |
        Returns immediately with 202 Accepted + submission_id. Backend processes
        asynchronously: validates image, calls the configured vision model
        (GLM-4V or Qwen-VL) for recognition and grading, annotates the original
        image, transcribes question text (English text / Math LaTeX), and updates
        the submission status. Frontend polls GET /api/submissions/{id} every 2s
        until status is completed/failed.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, subject, child_id]
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG/PNG exam photo (pre-compressed by frontend, max 2048px)
                subject:
                  type: string
                  enum: [english, math]
                child_id:
                  type: integer
                  description: Child ID from the parent's managed list
      responses:
        "202":
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/Unprocessable"

    get:
      operationId: listSubmissions
      tags: [Submissions]
      summary: List past submissions for the authenticated parent
      parameters:
        - name: child_id
          in: query
          required: false
          schema:
            type: integer
        - name: subject
          in: query
          required: false
          schema:
            type: string
            enum: [english, math]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SubmissionSummary"
                  total:
                    type: integer
  /api/submissions/{submission_id}:
    get:
      operationId: getSubmission
      tags: [Submissions]
      summary: Get submission detail (used for both polling and history)
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail (status may be pending/processing/completed/failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/submissions/{submission_id}/questions/{question_id}:
    patch:
      operationId: fixQuestionGrade
      tags: [Submissions]
      summary: Manually correct a question's grade
      description: |
        Allows parent to override the AI's grading for a specific question.
        Updates GradedQuestion.is_correct, sets is_manually_fixed=true,
        recalculates Submission.correct_count, and syncs ErrorQuestion.
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: integer
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_correct]
              properties:
                is_correct:
                  type: boolean
      responses:
        "200":
          description: Grade corrected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixQuestionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/images/{kind}/{filename}:
    get:
      operationId: serveImage
      tags: [Submissions]
      security: []
      summary: Serve an image file via a signed URL
      description: |
        Serves image files from the local filesystem. Authentication is via a
        signed URL (HMAC token + expiry in the query string), NOT the
        Authorization header, because <img>/<image> tags cannot send headers.
        Ownership is still verified through the submission_id embedded in the
        filename. Signed URLs are produced server-side and returned in API
        responses (image_url, thumbnail_url, etc.).
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [originals, annotated, thumbnails, questions, sheets]
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: "{submission_id}.jpg or {submission_id}_{qnum}.jpg or {uuid}.jpg"
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: HMAC signature token (base64url)
        - name: expires
          in: query
          required: true
          schema:
            type: integer
          description: Expiry as a Unix timestamp (default TTL 1 hour)
      responses:
        "200":
          description: Image file (JPEG binary)
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════════════
  # Error Collections
  # ═══════════════════════════════════════════════════════
...
components:
  schemas:
...
    SubmissionAccepted:
      type: object
      required: [submission_id, status]
      properties:
        submission_id:
          type: integer
        status:
          type: string
          enum: [pending]
          description: Always "pending" on acceptance
    SubmissionSummary:
      type: object
      required: [id, child_id, child_name, subject, status, created_at]
      properties:
        id:
          type: integer
        child_id:
          type: integer
        child_name:
          type: string
        subject:
          type: string
          enum: [english, math]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        score:
          type: object
          required: [correct, total]
          properties:
            correct:
              type: integer
            total:
              type: integer
          description: Present only when status is "completed"
        thumbnail_url:
          type: string
          format: uri
          description: Signed URL to a small preview of the annotated image (null until completed)
        created_at:
          type: string
          format: date-time
    Submission:
      allOf:
        - $ref: "#/components/schemas/SubmissionSummary"
        - type: object
          required: [original_image_url]
          properties:
            original_image_url:
              type: string
              format: uri
              description: Signed URL
            annotated_image_url:
              type: string
              format: uri
              description: Signed URL; present only when status is "completed"
            total_questions:
              type: integer
              description: Present only when status is "completed"
            correct_count:
              type: integer
              description: Present only when status is "completed"
            token_usage:
              type: object
              properties:
                provider:
                  type: string
                  example: qwen
                  description: Vision model provider (glm | qwen)
                model:
                  type: string
                  example: qwen-vl-max
                prompt_tokens:
                  type: integer
                completion_tokens:
                  type: integer
                total_tokens:
                  type: integer
              description: Vision model token consumption (present when completed). Legacy records may omit provider/model.
            questions:
              type: array
              items:
                $ref: "#/components/schemas/GradedQuestion"
              description: Present only when status is "completed"
            updated_at:
              type: string
              format: date-time

    # ── Grading ───────────────────────────────────────
    GradedQuestion:
      type: object
      required:
        - id
        - question_number
        - is_correct
        - question_type
        - is_manually_fixed
      properties:
        id:
          type: integer
        question_number:
          type: string
          minLength: 1
          maxLength: 20
          description: Question number as shown on paper (e.g., "1", "1a", "II-3")
          example: "1a"
        question_position:
          type: object
          nullable: true
          description: Bounding box as percentage of image dimensions
          properties:
            x:
              type: number
              description: Left edge (% of image width)
            "y":
              type: number
              description: Top edge (% of image height)
            w:
              type: number
              description: Width (% of image width)
            h:
              type: number
              description: Height (% of image height)
        question_image_path:
          type: string
          nullable: true
          description: Signed URL to the cropped question region image
        question_type:
          type: string
          enum: [choice, fill_blank, reading, composition, calculation, word_problem]
        is_correct:
          type: boolean
        question_text:
          type: string
          nullable: true
          description: Transcribed full question text (English questions).
        question_latex:
          type: string
          nullable: true
          description: Transcribed question in LaTeX (Math questions). Empty for English.
        solution_note:
          type: string
          nullable: true
          description: Solution approach for incorrect answers only
        error_category:
          type: string
          nullable: true
          enum: [grammar, vocabulary, spelling, logic, calculation, careless, comprehension]
        is_manually_fixed:
          type: boolean
          default: false
          description: True if parent overrode the AI grading

    # ── Manual Correction ────────────────────────────
    FixQuestionResponse:
      type: object
      required: [question, new_score]
      properties:
        question:
          $ref: "#/components/schemas/GradedQuestion"
        new_score:
          type: object
          required: [correct, total]
          properties:
            correct:
              type: integer
            total:
              type: integer

    # ── Error Collections ─────────────────────────────
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

  # ── Security Schemes ──────────────────────────────
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by POST /api/auth/login. The `sub` claim is the parent id
        (userId). Applies to all business endpoints.
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Future scheme — retained for migration reference only. NOT implemented
        in v2.

  # ── Responses ───────────────────────────────────────
  responses:
    BadRequest:
      description: |
        Invalid request — malformed input, unsupported format, or image that
        cannot be processed (unclear photo, wrong format, too large). Use for
        semantic validation failures.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "Image could not be processed. Ensure it is a clear photo of an exam paper."
    Unprocessable:
      description: |
        Request schema validation failure (FastAPI default 422). Used when a
        required field is missing or a field value violates its declared schema
        (wrong type, out-of-range enum). Do NOT use this for image-processing
        problems — those are 400.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "Field 'subject' must be one of: english, math."
    NotFound:
      description: Resource not found (also used for cross-owner access to avoid existence probing)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "Submission not found."
    Forbidden:
      description: |
        Forbidden — invalid or expired signature (signed image URL), or
        otherwise not permitted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "Invalid or expired image signature."